delivery

+++ src/main/java/com/sms/businesslogic/service/DeliverService.java
@@ -3,6 +3,7 @@ package com.sms.businesslogic.service;
 import com.sms.businesslogic.entity.Delivery;
 import com.sms.businesslogic.entity.Order;
 import com.sms.businesslogic.repository.DeliveryRepository;
+import com.sms.businesslogic.repository.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +18,8 @@ import java.util.Random;
 public class DeliverService {
     @Autowired
     private DeliveryRepository deliveryRepository;
+
+    private OrderRepository orderRepository;
 
     public List<Delivery> getAllDeliveries() {
         return deliveryRepository.findAll();
@@ -46,7 +49,11 @@ public class DeliverService {
         return "TRACK-" + orderId + "-" + randomNumber;
     }
 
-    public Delivery updateDelivery(Order order, String shippingAddress) {
+    public Delivery updateDelivery(Integer orderId, String shippingAddress) {
+
+        Order order = orderRepository.findById(orderId)
+                .orElseThrow(() -> new IllegalArgumentException("Order not found"));
+
         Delivery newDelivery = new Delivery();
         String trackingNumber = generateTrackingNumber(order.getOrderId());
 
@@ -55,6 +62,7 @@ public class DeliverService {
             newDelivery.setTrackingNo(trackingNumberAsInt);
         } catch (NumberFormatException e) {
             // Handle the case where the tracking number is not a valid integer
+            newDelivery.setTrackingNo(-1);
         }
         newDelivery.setDeliveryDate(calculateDeliveryDate());
         newDelivery.setDeliveryStatus("In Progress");
@@ -65,6 +73,10 @@ public class DeliverService {
     }
 
     public void deleteDelivery(Integer deliveryId) {
-        deliveryRepository.deleteById(deliveryId);
+        if (deliveryRepository.existsById(deliveryId)) {
+            deliveryRepository.deleteById(deliveryId);
+        } else {
+            throw new IllegalArgumentException("Delivery not found");
+        }
     }
 }
